fixed bugs with parking data and added pick image functionality to edit user profile

+++ parka/lib/data/dtos/parking/update_parking_dto.dart
@@ -4,7 +4,7 @@ class UpdateParkingDto {
   final String parkingId;
   int countParking;
   String parkingName;
-  double priceHours;
+  int priceHours;
   List<String> pictures = [];
   String mainPicture;
   String information;
